Removed allowedHosts and reportedHosts.
allowedHosts was pulled from Blok (https://github.com/mozilla/blok) for storing sites the user chose to disable TP for during the study. This is no longer a feature in the revised study design (https://tinyurl.com/y9zd6mnc).

ReportedHosts was pulled from Blok for storing sites/information when a user reported that the site didn’t look right or wasn’t working. This feature is not a part of the current study.

Also updated some comments per input from rhelmer, who was originally working on this study.