Store timeSaved in ms and convert to minutes in new tab view only
This prevents issues of dissimilar type (number vs string) and excessive decimal places when timeSaved is passed from the new-tab-variation.js frame script to Feature.jsm. timeSaved is stringified during message passing, so it needs to be reconverted back into a number in Feature.jsm for proper addition in this.state.totalTimeSaved.