Update pageAction badge with timeSaved
Because the frame script loads per page, not per tab, I need to reset the LOAD_START_TIME value every time there’s a new page being loaded in the same tab (or a page is refreshed). Otherwise the “load” event will calculate the difference between Date.now() and LOAD_START_TIME as an ever increasing quantity in the same tab.
My solution: Get rid of LOAD_START_TIME global variable in the frame script and use window.performance API to estimate page load time for each “load” event in the frame script.
This required some fiddling, as some pages trigger multiple load events, so we want to sum them for total time saved, but only show them in increasing order on the pageAction badge. We also needed to know when a page was refreshed or the user visited a different page in the tab, as we would want to reset the page load time value in memory and in the badge.

TODO: timeSaved in the badge is currently in ms, we need it in s and to add the units on the pageAction badge (s).