Initialize TP. Fixes #18. Reset TP at study end. Fixes #21.
This commit initializes the state of built-in Tracking Protection when the study starts based on the user's treatment branch. It also restores built-in tracking protection to its default setting when the study ends, except in the case that the study ends because the user proactively changed the state of Tracking Protection in about:preferences.