Add logs as telemetry placeholder for intro panel UI events.
There is now a console.log for the following intro panel UI events:
* When it is shown (add event listener "popupshown")
* When it is hidden (add event listener "popuphidden")
* How long it was shown for.
* When it is dismissed and why:
  - due to tab change (user switches to a different tab)
  - due to tab close (user closes the tab the panel is shown in)
  - due to page change in same tab
  - due to window close (user closes the window that the tab is shown in, "SSWindowClosing" event listener)

The event listeners added to the introPanel element do not have to be removed on feature.uninit() because the panel is a child element to the pageAction button, and the button is already removed in feature.uninit().
The window closing event listener ("SSWindowClosing") is removed in feature.uninit().