Add custom telemetry pings for intro panel UI
Any time the panel is shown, dismissed, or a button in the panel is clicked, a study-specific telemetry ping is sent with more details about the event.

The study now ends if the user disables TP from the intro panel's confirmation screen. There is a bug with this, see Issue #51.

Replaced instances of console.log with this.log.debug, so logs can be toggled ON/OFF with a setting in Config.jsm.