Remove modification of builtin TP UI.
The built-in TP UI pref 'privacy.trackingprotection.ui.enabled' doesn't actually control the TP UI in the chrome as initially thought; it only changes the UI for TP in 'about:preferences' from a set of 3 radio buttons to a single checkbox. We actually want the radio buttons enabled, which is the default setting.