Use sessionstore-windows-restored observer to startup feature. Fixes #73.
This may improve performance for the 'tspaint' and 'sessionrestore' TALOS tests, but from the original perspective outlined in #73, the previous observer (browser-delayed-startup-finished) worked just as well when Session Restore was enabled in 'about:preferences' (i.e. open Firefox with all previously opened tabs); I believe this is because the pages aren't loaded until the user focuses that particular tab. If the pages all automatically loaded at the same time on startup with Session Restore, this would make a difference.