Keep same .bind reference and balance listeners
After running this addon against mozilla-central on the try server, we saw some unit tests fail citing memory leakage; this change is an attempt to address that issue. See https://github.com/biancadanforth/tracking-protection-shield-study/issues/26#issuecomment-362155010 for more details.

In addition to ensuring each add/removeListener pair points to the same listener reference with 'bind', I added an eslint rule to check for balanced listeners to ensure that for every addEventListener, there's a corresponding removeEventListener.

In the process, I cleaned up the code in intro-panel.js and page-action-panel.js quite a bit.

I also updated the id of the addon in package.json, which required changing it where it appeared in other places (notably jar.mn and moz.build for building/running/testing in tree).