WIP - Fix addTabProgressListener leak on initial window.
This is the start of a fix for #88.

FIXED: The addTabsProgressListener now removes itself from the initial window if it closes BEFORE Feature.uninit (i.e. WindowWatcher.stop) is called. So now, we get mTabsProgressListeners.length down to 0 when that window is closed instead of 1.
Reason: We were previously not handling the onWindowClose event in the WindowWatcher observe function -- Are event listeners on windows automatically removed on window close? NO.

TODO:
* Listen for tab-switching-windows event (whatever that is) and remove the addTabsProgressListener from the old window and add an addTabsProgressListener for the new window. Completing this should make our example test, browser/base/content/test/permissions/browser_temporary_permissions_tabs.js, and likely many other tests, pass.
* addTabsProgressListener correctly to newly created windows.