Add addTabsProgressListeners when a new window opens.
Previously, the addTabsProgressListeners was not being added to new windows. Now it is.

As part of this, I edit the bind reference for the win._onWindowLoaded method and check for windowType instead of window location as a more idiomatic way to ensure the load listener is only added to DOMWindows (as opposed to DevTools windows or other kinds of windows).

Also made WindowWatcher a class exporting a single instance of that class for more consistency with my other modules.

Finally, since I've been building a local copy of Nightly for local/try server testing, I had to explicitly set an environment variable for my normal, non-dev build binary of Nightly, as 'npm run firefox' was otherwise opening the local build's binary instead. This is problematic right now since 'npm run firefox' installs the addon as a temporary addon, while the local build installs the addon as a system addon for testing.