Add local Firefox testing helper script. Fixes #91.
Replaced 'mach-test.sh' with 'test-in-tree.sh'. The latter now runs with the command 'npm run test', instead of being appended to 'npm run build'.

Right now the path to mozilla-unified and which test to run locally is hardcoded into 'test-in-tree.sh'; there should we an easy way to edit this for developers.