Override about:privatebrowsing homepage
The default home page in Private browsing mode is "about:privatebrowsing", rather than "about:newtab". This page is substantially different, including the fact that this page tells the user if built-in Tracking Protection is turned off, and gives them a toggle right on the page to turn it on.

If a user changes their Tracking Protection setting at any time during the study from what was assigned to them by their treatment branch, that ends the study.

This is important, because for the experimental branches, where users are told Tracking Protection is on and receive "fast" or "private" messaging, built-in Tracking Protection has to be set to off. This is because there is no API with built-in Tracking Protection for us to otherwise know the number of blocked resources per page, which is used in a pageAction badge and totaled in the new tab page for these users.

Thanks to a suggestion from Gijs, I was able to easily override the default home page in Private Browsing mode (about:privatebrowsing) with a modified version of the same page that omits the Tracking Protection toggle and an external link.