Ensure panels are dismissed properly on window close. Fixes #103.
TL;DR: remove the 'SSWindowClosing' event and just call its listener inside of the '_unloadCallback' method passed into 'WindowWatcher.jsm'.

---

A longer explanation:

Per MDN: The SSWindowClosing event is executed before a window is closed and the session store stops tracking it.

My interpretation of this is that the references to window objects that window may have had should still be in place when the event fires…so the bug is not related to the Weak Reference memory leak fixes as I originally thought.

However, the 'SSWindowClosing' event is not actually firing when I close the either the initial or second window anymore. Why?
* I know that the listener is successfully added to both the original and new window.
* There are no typos with respect to the bind reference being made and the method name…
* I did add a window close event listener to 'WindowWatcher.jsm'; turns out, that is firing now before the 'SSWindowClosing' event has a chance to fire. I.e. the Observer notification “domwindowclosed” in WindowWatcher fires BEFORE the SSWindowClosing in Feature. To confirm this, I commented out the WindowWatcher.jsm  _onWindowClosed callback and saw that SSWindowClosing does fire properly, and the panel gets properly dismissed when the observer unload callback _onWindowClosed is removed.

So, now that I have confirmed that if I remove the observer for 'domwindowclosed' in 'WindowWatcher.jsm', the panel gets properly dismissed via 'SSWindowClose' inside 'Feature.jsm', so what I really need is a way to ensure the panel gets dismissed properly by way of 'domwindowclosed'.

What I ended up doing: remove the 'SSWindowClosing' event and just call its listener inside of the '_unloadCallback' method passed into 'WindowWatcher.jsm'.