Show intro panel on pageAction click if badge is 0. Fixes #110.
Updated logic in 'handlePageActionButtonCommand' to open the intro panel if the user clicks the pageAction the first time they visit a page, even if the page hasn't blocked any resources. The intro panel can then be dismissed in any of the usual ways, now including by clicking the pageAction button again.

Also replaced 'this.shouldShowIntroPanel' with 'this.state.shouldShowIntroPanel', since this boolean is related to the state of the feature.